test(core): consolidate duplicate subaction, tool action, and activity suites (#27163)
